+++ webapp/tests/unit/i18n/utils/key-extractor.js
@@ -3,8 +3,8 @@
  * @tagline         Extract i18n keys from view and controller files
  * @description     Utility to extract translation keys from various source formats
  * @file            webapp/tests/unit/i18n/utils/key-extractor.js
- * @version         1.6.19
- * @release         2026-02-19
+ * @version         1.6.20
+ * @release         2026-02-20
  * @repository      https://github.com/jpulse-net/jpulse-framework
  * @author          Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
  * @copyright       2025 Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/