+++ webapp/tests/unit/i18n/utils/key-validator.js
@@ -3,8 +3,8 @@
  * @tagline         Validate translation keys against reference
  * @description     Utility to compare key sets and identify missing/extra keys
  * @file            webapp/tests/unit/i18n/utils/key-validator.js
- * @version         1.6.19
- * @release         2026-02-19
+ * @version         1.6.20
+ * @release         2026-02-20
  * @repository      https://github.com/jpulse-net/jpulse-framework
  * @author          Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
  * @copyright       2025 Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/